backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
GitPython
